Add paht to Discover Movies button
