+++ app/views/users/show.html.erb
@@ -1,6 +1,6 @@
 <h1><%= @user.name %>'s Dashboard</h1>
 
-<%= button_to "Discover Movies" %>
+<%= button_to "Discover Movies", user_parties_path(@user), method: :get%>
 
 <h3>Viewing Parties</h3>
 <% @parties.each do |party| %>
