+++ spec/features/users/show_spec.rb
@@ -22,15 +22,3 @@ describe 'User Dashboard Page' do
     expect(page).to_not have_content('Eternal Sunshine of the Spotless Mind')
   end
 end
-
-describe 'Dashboard: Discover Movies' do
-  before :each do
-    test_data
-    visit user_path(@user1)
-  end
-
-  it "when visiting a user dashboard and clicking the Discover Movies Button, it redirects to the specific user's discover page" do
-    click_button "Discover Movies"
-    expect(current_path).to eq()
-  end
-end
